build: removes intializeNotional calls footer data and adds lending message
